Adds soulful USCM SPF 50 sunscreen. (#2183)
# About the pull request

This adds a soulful SPF 50 stick of sunscreen. It comes with a
discounted cost to accommodate for the uselessness on night operations
as well as the long term risk of heavy metal poisoning.

This is my first item added to the game, and I did test it. Sprites
thanks to @thwompa

# Explain why it's good for the game

This is a simple, soulful addition that marines will surely spend points
on, and it diversifies the facepaint category as well as adding
potential for other, even more useless items to be added for lore.

# Testing Photographs and Procedure

I spawned it in, and it works. This is a
[peek](https://cdn.discordapp.com/attachments/948362058351656980/1058998160124563557/IMG_1537.jpg)
at what it looks like, thanks to the spriter.


</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl: Newyearnewme
add: Added sunscreen. Now you can enjoy those rays without risking
damage from UV rays!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->