fixes 2 typos and a piece of armor (#2235)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This is a super simple PR that fixes 2 typos and a piece of armor

-WWeyland-Yutani to Weyland-Yutani
-Genadier to Grenadier
-and removed artificial 'slowdown' from the map-spawning M4 synthetic
armor making it from slowdown = 1 to slowdown = 0.35

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Typos were bad to see, particularly Genadier

The M4 armor having a heavy armor slowdown makes no sense even if it's
on the map, because the thing has no armor to begin with. All its armor
values are none and it already has a natural slowdown of light armor.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->



# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
fix: removed unintended slowdown from the M4 light synthetic armor on
nightmare
spellcheck: fixes WWeyland-Yutani and Genadier
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->