fix corpse spawner on big red medbay nightmare insert (#2251)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

Fixes corpse spawner on nightmare insert

# Explain why it's good for the game

Corpse spawner is spawning live mobs which shouldn't be the case. The
was a result of the wy goon corpse update. The update changed one of the
corpse spawner to just an unknown default which shouldn't be the case.


# Testing Photographs and Procedure

![image](https://user-images.githubusercontent.com/103712112/211173925-e1c6acee-6428-437f-8ed1-3ff8bb433907.png)


![image](https://user-images.githubusercontent.com/103712112/211173930-0f9a35cd-9fa9-4b27-b54f-a1c98ad4b39a.png)


</details>


# Changelog

:cl: theflyingflail
fix: fixed broken spawner on big red nightmare insert
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->