Fix for do_after() and alpha check. (#2541)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Followup to #2456. In as much as reverting something and redoing it from
scratch can be called a followup.
I was going to go on entire _rant_ about how when you copypaste
somebody's code to retry making it, you should watch carefully whence
and whither you are copypasting (that `break` belonged in its own `if`
and not in the one inserted into the middle of it), how you should test
that your version actually works, how `TRUE` and `FALSE` are different
things and one thing is not another (this one should be obvious,
but...). Then I switched to going to go on a rant about some extreme
edge cases and how terrible Byond is at handling something about icons
and alphas. Then I went to actually read documentation and realised that
we have been trying to reinvent a bicycle and that it was not even going
to liftoff anyway.
When all you had to do was revert the (relevant part of)
[this](https://github.com/cmss13-devs/cmss13/commit/d523bb75462cc9367cc55ee8ba3f07e16c9d3acf).

# Explain why it's good for the game

Everything still broke from #2456, though. This has made a lot of people
very angry and been widely regarded as a bad move.

# Testing Photographs and Procedure
...and I still reserve the right to say this: _actually test that your
code actually works_.


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
fix: Timed actions are once again actually correctly timed.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->