Fixes Lakehouse Nightmare Insert Spawner (#2583)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# Fixes Lakehouse Nightmare Insert Spawner

Fixes the nightmare insert for LV-624 Survivor spawner. When it spawns
all survivors are supposed to spawn there as a last stand but that is
not the case. Sometimes only one survivor spawns and sometimes no-one
does.

# Explain why it's good for the game

The nightmare insert deserves some love when it spawns. Its a cool idea
that needs a quick fix and a touchup later down the line

:cl: theflyingflail
fix: fixed skylight nightmare insert survivor spawner
:cl:

<!-- Both :cl:'s are required for the changelog to work! -->