Colony Guns Part 3: Longarms Rework (#2334)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Updates and buffs the following weapons:

- Basira-Armstrong Rifle
- Model 12 Bolt-Action Rifle
- M37-17 pump shotgun
- Dragunov sniper rifle

### Basira-Armstrong Rifle

- The BA rifle has been reflavored as the removed 'ceremonial' ABR-40,
now a hunting-civilian version of the L42. It effectively has the same
stats as the L42, just don't take the stock off!
- Its magazines can only fit 12 bullets, but they still have the classic
L42 kick to them. The magazines are completely compatible between both
military and civilian gun types.
- Additionally, there are now holo-targeting magazines available for the
ABR-40, for hunting target capture purposes.
- - The Contractor ERT has a chance to spawn with a tacticool ABR-40
loadout, with three spare holotarget magazines on their webbing.

### Model 12 Bolt-Action Rifle

- Like its sprite implies, this is now the new Basira-Armstrong rifle.
- Its stats have recieved an overhaul and it can now hold its own
against a Xenomorph (or marine if you're CLF)
- Additionally, its bullets will push back (not stun) targets within
three meters, to increase viability for colony usage.

### M37-17 Pump Shotgun

- Did you know that the 10% damage mod this gun was supposed to have
didn't work? Now it does! And it deals 15% more damage to make up for
it.
- However, it can now be melted down.

### Dragunov Sniper Rifle

- The dragunov has been revamped into a DMR, needing no skill to fire,
dealing good damage, and having the same push-back feature the
bolt-action now has.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

All of these weapons are seen as, and are, a complete joke and ignored
in favor of chungillionaire among us sus impostor running around PBing
xenos with buckshot. Buffing them to be useful paves the way for
civilians to use civilian weaponry instead of chungillionarius.

### ABR-40 Hunting Rifle

This is intended as asset recycling. IIRC, Triiodine disliked the
existence of a ceremonial rifle and thought that was goofy, which is an
understandable opinion, but it pains me to see the gun's cool assets go
unused. I took the opportunity and made the lame and mediocre Basira
into this cool gun which marines and survivors will take a lot more
interest in than a Basira plinker.

While it does have L42 stats, colonists won't be able to locate any L42A
ammo on the colony, meaning they are limited to 12-round magazines.
Still, giving them a weapon like this is a great way to incentivize them
to step out of their chungus zone.

As for the contractor addition, these are supposed to be ex-USCM
mercenaries, who are trained in the usage of marine equipment. It makes
sense that they'd grab the civilian version of a primary marine gun and
tune it to their liking.

### Basira-Armstrong Bolt-Action Hunting Rifle

Chomp made this really cool backend for a bolt-action that tragically
ended up never being used, not really, aside from a half-hearted few
rifles being thrown in at Shiva's Snowball. Since these guns are so
weak, it's plain to see why nobody even looks at them twice. So I
changed that.

### M37-17 Pump Shotgun

Bugfix and a small damage buff to make up for it. It being unacidable
was lame anyways. Spread projectiles are still bugged and don't inherit
the base gun's damage mod, but that's out of scope.

### Dragunov Designated Marksman Rifle

This gun has been a joke since 2017, it's time to give it some love.
Still needs one-hand sprites but not my problem.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
balance: Updates and buffs the following weapons: Basira-Armstrong
Rifle, Model 12 Bolt-Action Rifle, M37-17 pump shotgun, Dragunov sniper
rifle
imageadd: The Basira rifle has been reflavored as the removed
'ceremonial' ABR-40, now a hunting-civilian version of the L42. It
effectively has the same stats as the L42, just don't take the stock
off!
add: Its magazines can only fit 12 bullets, but they still have the
classic L42 kick to them. The magazines are completely compatible
between both military and civilian gun types.
add:: Additionally, there are now holo-targeting magazines available for
the ABR-40, for hunting target capture purposes.
add: The Contractor ERT has a chance to spawn with a tacticool ABR-40
loadout, with three spare holotarget magazines on their webbing.
balance: Like the bolt-action rifle's sprite implies, this is now the
new Basira-Armstrong rifle.
add: Its stats have recieved an overhaul and it can now hold its own
against a Xenomorph (or marine if you're CLF)
add: Additionally, its bullets will push back (not stun) targets within
three meters, to increase viability for colony usage.
balance: Did you know that the 10% damage mod the M37-17 pump shotgun
was supposed to have didn't work? Now it does! And it deals 15% more
damage to make up for it.
del: However, it can now be melted down.
balance: The dragunov has been revamped into a DMR, needing no skill to
fire, dealing good damage, and having the same push-back feature the
bolt-action now has.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->