Unique 7-stage larva HUD update (#2616)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
Stages 0 and 1, and 4 and 5 of a larval infection shared the same icon
states in their HUD and their palettes were not universal, so I created
unique sprites for each larval stage.

Before: 

![larva2](https://user-images.githubusercontent.com/13158711/215290996-5edc5a52-b983-4e1f-9e74-b36b3191d3be.png)

After:

![image](https://user-images.githubusercontent.com/13158711/216842833-ecff3f6c-a742-4997-a157-b7db0dcef5db.png)

Finally, the Xeno Leader star received a small update so it actually
looks like a star. Left is before, right is after. It was also moved one
pixel to the right so it would not overlap on the Prime icon state.

![image](https://user-images.githubusercontent.com/13158711/215291039-b6c1bc3f-6ab6-4252-870b-21c86bbf5d4b.png)

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->
These are just a few quality of life changes. Plus, the star actually
looks like a star, now, so, that's a good thing, now!

# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
fix: Stage 1 and stage 4 of a larval infection will now properly be
distinguished from other stages in a Xenomorph's HUD.
tweak: All larva stages in the Xenomorph HUDs now each share a unique
sprite and have a uniform color palette.
tweak: The Xenomorph Leader star has been reshaped into a proper star
with added shading and moved a pixel to the right to not overlay any
other HUD icon states.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->