Taser Mode Refactor (#2482)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
Gave taser mode an action button. Uses the ID lock icon as it fits with
the idea of using or not using the wanted database.
Also changed how it's handled on the back end, giving it a proc rather
than replacing unique action with it.
Added an overlay on the taser itself that also shows what mode its
using.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes it a bit more functional, also clearer to see what you're using.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

![image](https://user-images.githubusercontent.com/41653574/214873035-b3e40cca-5537-40a8-82cd-cba935907abe.png)

![image](https://user-images.githubusercontent.com/41653574/214873052-a4af9446-7ce2-4590-a53a-45bdde2a4707.png)

![image](https://user-images.githubusercontent.com/41653574/214873067-ed6408ab-a1d0-40d2-81fb-c4668556b4bb.png)


</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
refactor: Changing mode on the taser now uses an action button and
unique proc rather than replacing unique action. Can still be triggered
with unique action.
imgadd: Added an overlay on the taser to denote what mode it's using.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->