Cat sprite revert, attempt 3 (#2551)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR is #2258 but without conflicts in the dmi files that github
refuses to resolve otherwise. The rest is copied and pasted from there:

This is essentially reopened #2020  , but now with extras.

For a full breakdown why the sprites deserve a revert, please consult my
treatise found in #2020 TLDR: The new sprites are cartoony and
anatomically incorrect and clash with CM's fairly realistic art style.
Also they turned Jones into a tabby, unacceptable.

In addition to reverting the sprites, this does the following:

- Adds an inhand sprite for Jones/calicos, we were missing that for some
reason.
- Organizes the cat holders so they make sense.
- For consistency's sake adds a generic black cat mob type along
Runtime.
- Removes being able to gib cats using ATM machines, apparently added as
a funny meme reference. Killing Jones indiscriminately is already an
issue to the point that it was sanctioned OOCly, having a meme way of
doing that is not helping anything. Also, it is incredibly upsetting to
know we even had that in the code in the first place.


<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Sprite consistency good.
Code consistency good.
Meme mechanics that are essentially ooc rule breaks bad.

Also _all_ cats are best cats. 

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:Silencer_pl
imageadd: Older, more anatomically correct and less cartoony cat sprites
have been readded. Jones is a calico again!
del: Cats can no longer be gibbed in ATMs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>