Opens up 2 spots in Trijent's sand temple area. (#2640)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
This PR fixes the issue of an impassable rock blocking two sections of
the Trijent map's sand temple that marines _must_ C4, and opens up the
area to its direct south. I also added a little fun description for the
smaller boulder I turned the large 2x2 into referencing how in the way
it is.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Here's the before-

![image](https://user-images.githubusercontent.com/70115628/218394201-9c65c885-730f-40e6-a40e-abc3c931db85.png)
Note how inconvenient this is for both sides.
Here's the after-

![image](https://user-images.githubusercontent.com/70115628/218394339-85bcabfd-7ece-40bf-a62e-179b898068c1.png)

This is a pretty big improvement, and it makes this area easier for
marines and xenos to use and fight in. I've seen this artificially
extend hive assaults which doesn't seem fun for either side.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>


![image](https://user-images.githubusercontent.com/70115628/218395888-476274c6-eb0b-45f7-9955-b3359c69bcd3.png)
more movement to fight changes everything


</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
maptweak: opened up 2 spots in Trijent's sand temple area
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->