Seegson Cargo Container Recolor (#2681)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Changed the green Seegson cargo container to a yellow cargo container as
Seegson's company color is yellow.

# Explain why it's good for the game

Its on brand and to be more yellow. Besides we need more variety in
cargo container colors, almost all of them are red, blue, and green!

# Testing Photographs and Procedure

<details>

![Screenshot 2023-02-19
012958](https://user-images.githubusercontent.com/108848001/219879909-22db03b3-d76b-4011-b671-68719c689dc7.png)

</details>

# Changelog

:cl: Tropical_Flak

imageadd: Changed the Seegson cargo container to be yellow instead of
green.

/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>